通过model_worker支持智普AI在线接口 (#1300)
* 更新shutdown_all.sh:更新mac设备的额外说明

* update VERSION

* Add files via upload

* Update README.md

* search_engine_chat bug

* 修复 milvus 作为默认数据库初始化异常和调用search方法异常

* Add files via upload

* Update README.md

* 修复搜索引擎聊天接口 (#1228)

* 修复搜索引擎聊天接口

* 修复知识库聊天接口

* 恢复

* 调整KBService 接口函数do_search 并适配pg

* 修正 milvus 更改错误

* update readme about startup.py

* 在search_engine_chat中检查Bing KEY，并更新tests

* update server_config.py.example: 增加多卡启动的说明

* change default search_engine option to 'duckduckgo'

* update readme.md, shutdown_all.sh: 在Linux上使用ctrl+C退出可能会由于linux的多进程机制导致multiprocessing遗留孤儿进程

* update VERSION

* Update README.md

* update requirements.txt

* update requirements

* update requirements

* update VERSION

* [BUG] 修复csv文件读取后，单行数据被分成多段。

* update INSTALL.md (#1250)


* Update requirements.txt

* fix #1224: failed to load json and pdf files (#1252)


* close #1172: 给webui_page/utils添加一些log信息，方便定位错误

* 修复：重建知识库时页面未实时显示进度

* 修复重建知识库界面无反应的问题，添加一些log信息 (#1258)

* close #1172: 给webui_page/utils添加一些log信息，方便定位错误

* 修复：重建知识库时页面未实时显示进度

---------


* 适配score_threshold

* skip model_worker running when using online model api such as chatgpt

* 修复使用chatgpt等在线API服务时模型加载错误 (#1271)

* close #1172: 给webui_page/utils添加一些log信息，方便定位错误

* 修复：重建知识库时页面未实时显示进度

* skip model_worker running when using online model api such as chatgpt

---------


* update format

* 修改知识库管理相关内容:
1.KnowledgeFileModel增加3个字段：file_mtime(文件修改时间),file_size(文件大小),custom_docs(是否使用自定义docs)。为后面比对上传文件做准备。
2.给所有String字段加上长度，防止mysql建表错误(pr#1177)
3.统一[faiss/milvus/pgvector]_kb_service.add_doc接口，使其支持自定义docs
4.为faiss_kb_service增加一些方法，便于调用
5.为KnowledgeFile增加一些方法，便于获取文件信息，缓存file2text的结果。

* 修复/chat/fastchat无法流式输出的问题

* 新增功能：
1、KnowledgeFileModel增加"docs_count"字段，代表该文件加载到向量库中的Document数量，并在WEBUI中进行展示。
2、重建知识库`python init_database.py --recreate-vs`支持多线程。

其它：
统一代码中知识库相关函数用词：file代表一个文件名称或路径，doc代表langchain加载后的Document。部分与API接口有关或含义重叠的函数暂未修改。

* 统一XX_kb_service.add_doc/do_add_doc接口，不再需要embeddings参数

* 增加数据库字段，重建知识库使用多线程 (#1280)

* close #1172: 给webui_page/utils添加一些log信息，方便定位错误

* 修复：重建知识库时页面未实时显示进度

* skip model_worker running when using online model api such as chatgpt

* 修改知识库管理相关内容:
1.KnowledgeFileModel增加3个字段：file_mtime(文件修改时间),file_size(文件大小),custom_docs(是否使用自定义docs)。为后面比对上传文件做准备。
2.给所有String字段加上长度，防止mysql建表错误(pr#1177)
3.统一[faiss/milvus/pgvector]_kb_service.add_doc接口，使其支持自定义docs
4.为faiss_kb_service增加一些方法，便于调用
5.为KnowledgeFile增加一些方法，便于获取文件信息，缓存file2text的结果。

* 修复/chat/fastchat无法流式输出的问题

* 新增功能：
1、KnowledgeFileModel增加"docs_count"字段，代表该文件加载到向量库中的Document数量，并在WEBUI中进行展示。
2、重建知识库`python init_database.py --recreate-vs`支持多线程。

其它：
统一代码中知识库相关函数用词：file代表一个文件名称或路径，doc代表langchain加载后的Document。部分与API接口有关或含义重叠的函数暂未修改。

---------


* 运行startup.py时，如果不加参数直接显示配置和帮助信息后退出

* 运行startup.py时，如果不加参数直接显示配置和帮助信息后退出 (#1284)

* 统一XX_kb_service.add_doc/do_add_doc接口，不再需要embeddings参数

* 运行startup.py时，如果不加参数直接显示配置和帮助信息后退出

---------


* update server_config.py:
- 在model_config中增加HISTORY_LEN配置参数
- 将server_config中helper function移动到server.utils中
- 统一set_httpx_timeout的定义和调用

* update webui.py:
应用model_config中的配置项：HISTORY_LEN,VECTOR_SEARCH_TOP_K,SEARCH_ENGINE_TOP_K

* 优化server_config配置项 (#1293)

* update server_config.py:
- 在model_config中增加HISTORY_LEN配置参数
- 将server_config中helper function移动到server.utils中
- 统一set_httpx_timeout的定义和调用

* update webui.py:
应用model_config中的配置项：HISTORY_LEN,VECTOR_SEARCH_TOP_K,SEARCH_ENGINE_TOP_K

---------


* fix bug

---------

+++ server/knowledge_base/kb_service/milvus_kb_service.py
@@ -1,12 +1,16 @@
 from typing import List
 
+import numpy as np
+from faiss import normalize_L2
 from langchain.embeddings.base import Embeddings
 from langchain.schema import Document
 from langchain.vectorstores import Milvus
+from sklearn.preprocessing import normalize
 
 from configs.model_config import SCORE_THRESHOLD, kbs_config
 
-from server.knowledge_base.kb_service.base import KBService, SupportedVSType
+from server.knowledge_base.kb_service.base import KBService, SupportedVSType, EmbeddingsFunAdapter, \
+    score_threshold_process
 from server.knowledge_base.utils import KnowledgeFile
 
 
@@ -36,32 +40,22 @@ class MilvusKBService(KBService):
     def _load_milvus(self, embeddings: Embeddings = None):
         if embeddings is None:
             embeddings = self._load_embeddings()
-        self.milvus = Milvus(embedding_function=embeddings,
+        self.milvus = Milvus(embedding_function=EmbeddingsFunAdapter(embeddings),
                              collection_name=self.kb_name, connection_args=kbs_config.get("milvus"))
 
     def do_init(self):
         self._load_milvus()
 
     def do_drop_kb(self):
-        self.milvus.col.drop()
+        if self.milvus.col:
+            self.milvus.col.drop()
 
-    def do_search(self, query: str, top_k: int, embeddings: Embeddings):
-        # todo: support score threshold
-        self._load_milvus(embeddings=embeddings)
-        return self.milvus.similarity_search_with_score(query, top_k)
+    def do_search(self, query: str, top_k: int, score_threshold: float, embeddings: Embeddings):
+        self._load_milvus(embeddings=EmbeddingsFunAdapter(embeddings))
+        return score_threshold_process(score_threshold, top_k, self.milvus.similarity_search_with_score(query, top_k))
 
-    def add_doc(self, kb_file: KnowledgeFile, **kwargs):
-        """
-        向知识库添加文件
-        """
-        docs = kb_file.file2text()
+    def do_add_doc(self, docs: List[Document], **kwargs):
         self.milvus.add_documents(docs)
-        from server.db.repository.knowledge_file_repository import add_doc_to_db
-        status = add_doc_to_db(kb_file)
-        return status
-
-    def do_add_doc(self, docs: List[Document], embeddings: Embeddings, **kwargs):
-        pass
 
     def do_delete_doc(self, kb_file: KnowledgeFile, **kwargs):
         filepath = kb_file.filepath.replace('\\', '\\\\')
@@ -70,7 +64,8 @@ class MilvusKBService(KBService):
         self.milvus.col.delete(expr=f'pk in {delete_list}')
 
     def do_clear_vs(self):
-        self.milvus.col.drop()
+        if self.milvus.col:
+            self.milvus.col.drop()
 
 
 if __name__ == '__main__':
@@ -82,4 +77,4 @@ if __name__ == '__main__':
     milvusService.add_doc(KnowledgeFile("README.md", "test"))
     milvusService.delete_doc(KnowledgeFile("README.md", "test"))
     milvusService.do_drop_kb()
-    print(milvusService.search_docs("测试"))
+    print(milvusService.search_docs("如何启动api服务"))
